controls types, folder structure refactoring
