--- frontend/src/shared/ui/AuthPage/types.ts
@@ -1,8 +1,0 @@
-export interface IInputProps {
-  id: string;
-  label: string;
-  required: boolean;
-  placeholder?: string;
-  error?: string;
-  [K: string]: unknown;
-}
