+++ frontend/src/shared/ui/controls/inputs/PasswordInput/PasswordInput.tsx
@@ -3,16 +3,17 @@ import { FC, useState } from "react";
 import { SrOnlyLabel } from "@/shared/styles/globalStyledComponents";
 import { EyeNoneIcon, EyeOpenIcon } from "@radix-ui/react-icons";
 
-import * as S from "./AuthInput.styles";
-import { IInputProps } from "./types";
+import * as S from "../input.styles";
+import { IInputProps } from "../input.types";
 
-export const PasswordInput: FC<IInputProps> = ({ id, label, required, placeholder, error, ...rest }) => {
+export const PasswordInput: FC<IInputProps> = ({ id, label, required, placeholder, error, field }) => {
   const [isPasswordRevealed, setIsPasswordRevealed] = useState<boolean>(false);
+
   return (
     <S.FormInputContainer>
       <SrOnlyLabel htmlFor={id}>{label}</SrOnlyLabel>
       <S.PasswordInput
-        {...rest}
+        {...field}
         required={required}
         id={id}
         type={isPasswordRevealed ? "text" : "password"}
@@ -21,6 +22,7 @@ export const PasswordInput: FC<IInputProps> = ({ id, label, required, placeholde
         aria-invalid={!!error}
         aria-errormessage={`${id}_error`}
       />
+
       <S.PasswordRevealToggle>
         {isPasswordRevealed ? (
           <EyeNoneIcon role="button" aria-label="reveal password symbols" onClick={() => setIsPasswordRevealed(false)} />
@@ -28,6 +30,7 @@ export const PasswordInput: FC<IInputProps> = ({ id, label, required, placeholde
           <EyeOpenIcon role="button" aria-label="hide password symbols" onClick={() => setIsPasswordRevealed(true)} />
         )}
       </S.PasswordRevealToggle>
+
       {error ? <S.FormInputError id={`${id}_error`}>{error}</S.FormInputError> : null}
     </S.FormInputContainer>
   );
