+++ frontend/src/shared/ui/controls/inputs/PasswordInput/index.ts
@@ -1,0 +1,1 @@
+export { PasswordInput } from './PasswordInput';
