+++ frontend/src/shared/ui/controls/inputs/TextInput/TextInput.tsx
@@ -2,15 +2,15 @@ import { FC } from "react";
 
 import { SrOnlyLabel } from "@/shared/styles/globalStyledComponents";
 
-import * as S from "./AuthInput.styles";
-import { IInputProps } from "./types";
+import * as S from "../input.styles";
+import { IInputProps } from "../input.types";
 
-export const TextInput: FC<IInputProps> = ({ id, label, required, placeholder, error, ...rest }) => {
+export const TextInput: FC<IInputProps> = ({ id, label, required, placeholder, error, field }) => {
   return (
     <S.FormInputContainer>
       <SrOnlyLabel htmlFor="signup_email">Email</SrOnlyLabel>
       <S.FormInput
-        {...rest}
+        {...field}
         required={required}
         id={id}
         type="text"
