+++ frontend/src/shared/ui/controls/inputs/TextInput/index.ts
@@ -1,0 +1,1 @@
+export { TextInput } from './TextInput';
