+++ frontend/src/shared/ui/controls/inputs/index.ts
@@ -1,0 +1,2 @@
+export { PasswordInput } from './PasswordInput';
+export { TextInput } from './TextInput';
