+++ frontend/src/shared/ui/controls/inputs/input.types.ts
@@ -1,0 +1,12 @@
+import {  UseControllerReturn } from "react-hook-form";
+
+type RHFFieldProps = Partial<UseControllerReturn<any>['field']>;
+
+export interface IInputProps {
+  id: string;
+  label: string;
+  required: boolean;
+  placeholder?: string;
+  error?: string;
+  field?: RHFFieldProps;
+}
