+++ frontend/src/shared/ui/layouts/AuthPageLayout/AuthPageLayout.styles.tsx
