+++ frontend/src/shared/ui/layouts/AuthPageLayout/index.ts
@@ -1,0 +1,1 @@
+export { AuthPageLayout } from './AuthPageLayout';
