+++ frontend/src/shared/ui/layouts/MainLayout/MainLayout.styles.ts
