+++ frontend/src/shared/ui/layouts/MainLayout/MainLayout.tsx
