+++ frontend/src/shared/ui/layouts/MainLayout/index.ts
