+++ frontend/src/shared/ui/layouts/index.ts
@@ -1,0 +1,2 @@
+export { MainLayout } from './MainLayout';
+export { AuthPageLayout } from './AuthPageLayout';
